+++ src/lib/converter/plugins/CategoryPlugin.ts
@@ -249,7 +249,7 @@ export class CategoryPlugin extends ConverterComponent {
             }
         }
         if (aWeight === bWeight) {
-            return a.title > b.title ? 1 : -1;
+            return a.title.localeCompare(b.title);
         }
         return aWeight - bWeight;
     }